Updated Learnset of Gen2 to Indigo Disc (#469)
* Updated Learnset of Gen1 to Indigo Disc

Not many changes here for how much time it takes to comb through, but here it is.

Only notable change is Power-Up Punch got removed this gen so it had to be removed from Hitmonchan's learnset. Are we okay with this change? If so, this is good to go for Gen 1.

* Updated Learnset of Gen2 to Indigo Disc

Notable changes:

Porygon 2 loses Magic Coat with no replacement.

Ho-oh loses Burn Up but gains Overheat.

The rest seems better or neutral.