Update Learnsets for Gen3 to Indigo Disk (#571)
* Updated Learnset of Gen1 to Indigo Disc

Not many changes here for how much time it takes to comb through, but here it is.

Only notable change is Power-Up Punch got removed this gen so it had to be removed from Hitmonchan's learnset. Are we okay with this change? If so, this is good to go for Gen 1.

* Updated Learnset of Gen2 to Indigo Disc

Notable changes:

Porygon 2 loses Magic Coat with no replacement.

Ho-oh loses Burn Up but gains Overheat.

The rest seems better or neutral.

* Update Learnsets for Gen3 to Indigo Disk

A good chunk of these changes is just reordering the moves to match bulbapedia so that it's easier to see what changed in the future. Otherwise, it's really just replacing moves that don't exist anymore.

Also, Zangoose got a ton of new moves that work with a move relearner, but not by level-up. Weird, but good if you like Zangoose.

* Fix Duplicates