Update Item Check
Updated item check for when applying item to a Pokemon

Used similar code to when generating map for After Battle items list. Gets all item triggers for the Pokemon, then maps it and flattens it down to just a list of items the pokemon can use to change form. Then finally it checks if the current item exists in the list.