Implemented "Triple Axel" and "Triple Kick" increasing damage functionality
Triple Axel and Triple Kick have a functionality where their damage is multiplied by 2 on the second hit and 3 on the third hit.

A comment indicates that this functionality would theoretically be added in the MultiHitAddr class for MultiHitType._3_INCR, but it seems that this only really serves to provide a "hitTimes" value.

It seemed more reasonable to add a VariablePowerAttr for this kind of move that increases the power based on number of hits left as VariablePowerAttr is applied for every hit.