Implemented Triple Kick and Triple Axel properly
Added a proposed implementation for cyclically incrementing power in the moves Triple Kick and Triple Axel.

This implementation is based on MultiHitIncrementerThree which extends VariablePowerAttr.

This should cause each hit of these abilities to do 1, 2, or 3 times its base power on each hit.

This is repeated for each multi lens the Pokemon is holding as well.

For example, a Pokemon holding 1 multi lens using Triple Axel would have each hit of Triple Axel have a power, respectively, of: 20, 40, 60, 20, 40 60.