Remove enemy attack pre-calc
Due to the bugs it was causing, I've reverted the change I made where the enemy performs its attacks before you.

As a result, this breaks the following features that rely on this:
- Enemy attack prediction
- Catch prediction (I'm so sorry) (this has been reverted to showing % chance of success until I rework it)

In addition, if having a shiny breaks the shop, the minimum luck value that breaks the shop is shown on the charm.