Fix #5159: Harsh Sun, Heavy Rain, Delta Stream Messages Incorrect
The abilities Desolate Land, Primordial Sea and Delta Stream display
the normal weather effect messages instead of special ones (e.g.,
"the sunlight got hot" instead of
"The sunlight turned extremely harsh!").

Another issue is that if one of these special weathers is active
and the opponent uses a move that is ineffective in this weather,
it doesn't display a special message but the normal one
(e.g., using a Water-type move in Harsh Sunlight).

- Updated `weather.json` with the correct texts in:
  - `heavyRainStartMessage`
  - `harshSunStartMessage`
  - `strongWindsStartMessage`
- Added new "effect" messages to fix the fail message bug:
  - `heavyRainEffectMessage`
  - `harshSunEffectMessage`
  - `defaultEffectMessage`
- Added `getWeatherBlockMessage(WeatherType)` in `weather.ts`
  to return the correct message when a move fails
  due to Primordial Sea or Desolate Land.
- Called `getWeatherBlockMessage()` in `useMove()` (line 385)
  to update `failedText` using `failedDueToWeather`.
- Added `getWeatherType()` in `arena.ts` as a getter
  for `WeatherType`, since it wasn’t implemented yet.