Fix #5085: Moves dont play a No Effect Message Against Immune Type When using non-volatile status move like: Will-O-Wisp, Thunder Wave, Toxic, or Poison Gas against a Pokémon whose type is immune to that Status condition, no "It doesn't affect" message plays.
My proposed fixes:
In move.ts:
Updated isTypeImmune() function to receive another parameter
"move: Move" and to include the cases in which the status moves listed
in the bug are being used against the corresponding immune type
pokemon.
In pokemon.ts:
Updated the function call of "isTypeImmune()" to correspond with
the newly added parameter.