Fix #5085  Moves dont play a No Effect Message Against Immune Type
When using non-volatile status move like: Will-O-Wisp, Thunder
Wave, Toxic, or Poison Gas against a Pokémon whose type is immune
to that Status condition, no "It doesn't affect" message plays.

My proposed fixes:
In move.ts:
Removed the queue of the message ""XXXX is protected by Safeguard!"
in the class: "StatusEffectAttr" due to the Pokemon being
safeguarded when a status move is used and added a call to the
function canSetStatus().

In pokemon.ts:
Added the function "messageIsImmune(quiet: boolean)" which displays
the message "It doesn't affect XXXX" if the quiet parameter is set
to false.
Updated the "canSetStatus()" function so that when it returns false
(meaning the status couldn't be applied) it calls the
"messageIsImmune()" function to display the message correctly.
Also included the queue of the message removed from the move.ts
file allowing the centralization of the messages associated with
status moves.