Fix #5085  Moves dont play a No Effect Message Against Immune Type
When using non-volatile status move like: Will-O-Wisp, Thunder
Wave, Toxic, or Poison Gas against a Pokémon whose type is immune
to that Status condition, no "It doesn't affect" message plays.

My proposed fixes:
In move.ts:
Removed the else condition in the apply() function in StatusEffectAttr
class.

In pokemon.ts:
Added the function "messageIsImmune(quiet: boolean, effect:
StatusEffect)" which displays the message "It doesn't affect XXXX"
or the " XXXX is already affected by  YYY" message, depending if the
effect trying to be set is already affecting the pokemon or if the
pokemon is immune to the status.
Updated the "canSetStatus()" function so that when it returns false
(meaning the status couldn't be applied) it calls the
"messageIsImmune()" function to display the message correctly.