Show battle end phase in wild battle if the last pokemon fled.
I don't know about you guys but if my enemy fled the field I'd count that as a win.

Simply skipping to next battle is really infuriating when the one pokemon with this ability has a really strong priority move. Especially when it shows up right before an important fight.