Fixes Issue #6369 where Poison Puppeteer didn't work with Toxic Chain.
I added a conditional block within the current if statement to check if the thing causing the poison was in fact a pokemon and not from another source. If it was then applyAbAttrs("ConfusionOnStatusEffectAbAttr", ...) is called where within that, the pokemon is confirmed to have the ability/passive Poison Puppeteer, and if it does the opposing pokemon becomes confused.